+++ intellij-ext/modules/core/src/main/kotlin/com/metalbear/mirrord/MirrordExecDialog.kt
@@ -14,9 +14,10 @@ import javax.swing.event.DocumentListener
 object MirrordExecDialog {
     private const val dialogHeading: String = "mirrord"
     private const val targetLabel = "Select Target"
+    private const val configLabel = "Select Config"
 
-    fun selectTargetDialog(targets: List<String>): String? {
-        val jbTargets = targets.sorted().asJBList()
+    private fun selectionDialog(items: List<String>, label: String): String? {
+        val jbTargets = items.sorted().asJBList()
         val searchField = JTextField()
         searchField.document.addDocumentListener(object : DocumentListener {
             override fun insertUpdate(e: DocumentEvent) = updateList()
@@ -25,14 +26,14 @@ object MirrordExecDialog {
 
             private fun updateList() {
                 val searchTerm = searchField.text
-                val filteredTargets = targets.filter { it.contains(searchTerm, true) }.sorted()
+                val filteredTargets = items.filter { it.contains(searchTerm, true) }.sorted()
                 jbTargets.setListData(filteredTargets.toTypedArray())
             }
 
         })
         val result = DialogBuilder().apply {
             setCenterPanel(JPanel(BorderLayout()).apply {
-                add(createSelectionDialog(targetLabel, jbTargets, searchField), BorderLayout.WEST)
+                add(createSelectionDialog(label, jbTargets, searchField), BorderLayout.WEST)
             })
             setTitle(dialogHeading)
         }.show()
@@ -40,6 +41,14 @@ object MirrordExecDialog {
             return jbTargets.selectedValue
         }
         return null
+    }
+
+    fun selectTargetDialog(targets: List<String>): String? {
+        return selectionDialog(targets, targetLabel)
+    }
+
+    fun selectConfigDialog(configs: List<String>): String? {
+        return selectionDialog(configs, configLabel)
     }
 
     private fun List<String>.asJBList() = JBList(this).apply {